+++ weaver/processes/wps_package.py
@@ -143,25 +143,21 @@ if TYPE_CHECKING:
 
     from weaver.datatype import Authentication, Job
     from weaver.processes.constants import IO_Select_Type
-    from weaver.processes.convert import (
-        ANY_IO_Type,
-        CWL_Input_Type,
-        JSON_IO_Type,
-        PKG_IO_Type,
-        WPS_Input_Type,
-        WPS_Output_Type
-    )
+    from weaver.processes.convert import ANY_IO_Type, JSON_IO_Type, PKG_IO_Type, WPS_Input_Type, WPS_Output_Type
     from weaver.status import AnyStatusType
     from weaver.typedefs import (
         AnyHeadersContainer,
         AnyValueType,
         CWL,
         CWL_AnyRequirements,
+        CWL_Input_Type,
         CWL_IO_ComplexType,
+        CWL_IO_Type,
         CWL_Requirement,
         CWL_RequirementNames,
         CWL_RequirementsDict,
         CWL_Results,
+        CWL_SchemaNames,
         CWL_ToolPathObject,
         CWL_WorkflowStepPackage,
         CWL_WorkflowStepPackageMap,
@@ -1542,6 +1538,48 @@ class WpsPackage(Process):
             if self.package.get("baseCommand") == "python":
                 self.package["baseCommand"] = os.path.join(active_python_path, "python")
 
+    def update_cwl_schema_names(self):
+        # type: () -> None
+        """
+        Detect duplicate :term:`CWL` schema types not referred by name to provide one and avoid resolution failure.
+
+        Doing this resolution avoids reused definitions being considered as "conflicts" because of missing ``name``.
+        To avoid introducing a real conflict, names are injected only under corresponding :term:`CWL` I/O by ID.
+        The most common type of definition resolve this way is when :term:`CWL` ``Enum`` is reused for single and
+        array-based definitions simultaneously.
+
+        .. seealso::
+            - :func:`weaver.processes.convert.resolve_cwl_io_type_schema`
+            - :meth:`weaver.processes.wps_package.WpsPackage.make_inputs`
+
+        .. fixme::
+            Workaround for https://github.com/common-workflow-language/cwltool/issues/1908.
+        """
+        for io_select in ["inputs", "outputs"]:
+            if isinstance(self.package[io_select], dict):
+                io_items = self.package[io_select]  # type: Dict[str, CWL_IO_Type]
+            else:
+                io_items = {item["id"]: item for item in self.package[io_select]}  # type: Dict[str, CWL_IO_Type]
+            for io_name, io_def in io_items.items():
+                if isinstance(io_def["type"], list):
+                    item_enum = None
+                    array_enum = None
+                    for io_item in io_def["type"]:
+                        if not isinstance(io_item, dict):
+                            continue
+                        if io_item.get("type") == "enum":
+                            item_enum = io_item
+                            continue
+                        if io_item.get("type") != "array":
+                            continue
+                        if not isinstance(io_item.get("items", {}), dict):
+                            continue
+                        if io_item["items"].get("type") == "enum":
+                            array_enum = io_item["items"]
+                    # only apply the name reference if not already provided (eg: explicit name defined in original CWL)
+                    if item_enum and array_enum and item_enum == array_enum and "name" not in item_enum:
+                        item_enum["name"] = array_enum["name"] = f"{io_name}{uuid.uuid4()}"
+
     def update_effective_user(self):
         # type: () -> None
         """
@@ -1732,6 +1770,7 @@ class WpsPackage(Process):
 
             self.update_effective_user()
             self.update_requirements()
+            self.update_cwl_schema_names()
 
             runtime_params = self.setup_runtime()
             self.logger.debug("Using cwltool.RuntimeContext args:\n%s", json.dumps(runtime_params, indent=2))
@@ -1767,7 +1806,8 @@ class WpsPackage(Process):
                                                                                 eoimage_data_sources,
                                                                                 accept_mime_types,
                                                                                 settings=self.settings)
-                cwl_inputs = self.make_inputs(request.inputs, cwl_inputs_info)
+                cwl_schema_refs = package_inst.t.names.names
+                cwl_inputs = self.make_inputs(request.inputs, cwl_inputs_info, cwl_schema_refs)
                 self.update_status("Convert package inputs done.", PACKAGE_PROGRESS_CONVERT_INPUT, Status.RUNNING)
             except PackageException as exc:
                 raise self.exception_message(type(exc), None, str(exc))  # re-raise as is, but with extra log entry
@@ -1838,6 +1878,7 @@ class WpsPackage(Process):
     def make_inputs(self,
                     wps_inputs,         # type: Dict[str, Deque[WPS_Input_Type]]
                     cwl_inputs_info,    # type: Dict[str, CWL_Input_Type]
+                    cwl_schema_names,   # type: CWL_SchemaNames
                     ):                  # type: (...) -> Dict[str, ValueType]
         """
         Converts :term:`WPS` input values to corresponding :term:`CWL` input values for processing by the package.
@@ -1848,6 +1889,7 @@ class WpsPackage(Process):
 
         :param wps_inputs: Actual :term:`WPS` inputs parsed from execution request.
         :param cwl_inputs_info: Expected CWL input definitions for mapping.
+        :param cwl_schema_names: Mapping of CWL type schema references to resolve 'type: <ref>' if used in a definition.
         :return: :term:`CWL` input values.
         """
         cwl_inputs = {}
@@ -1859,7 +1901,7 @@ class WpsPackage(Process):
             # process single occurrences
             input_i = input_occurs[0]
             # handle as reference/data
-            io_def = get_cwl_io_type(cwl_inputs_info[input_id])
+            io_def = get_cwl_io_type(cwl_inputs_info[input_id], cwl_schema_names=cwl_schema_names)
             if isinstance(input_i, ComplexInput) or io_def.type in PACKAGE_COMPLEX_TYPES:
                 # extend array data that allow max_occur > 1
                 # drop invalid inputs returned as None